requirements.txt
Flask
flask-cors
Flask-SQLAlchemy
PyMySQL
opencv-python
numpy
matplotlib
scikit-image
pandas
mediapipe==0.10.13
scipy
pytest
